LPD-75170 refactor: rename ai-hub-rest-test to ai-hub-nexus-rest-test
