+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay AI Hub Nexus REST Test
+Bundle-SymbolicName: com.liferay.ai.hub.nexus.rest.test
+Bundle-Version: 1.0.0
+-includeresource:\
+	com.liferay.ai.hub.nexus.rest.client.jar=com.liferay.ai.hub.nexus.rest.client-*.jar;lib:=true,\
+	com.liferay.headless.batch.engine.client.jar=com.liferay.headless.batch.engine.client-*.jar;lib:=true,\
+	@jsonassert-[0-9.]*.jar
