+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/build.gradle
@@ -23,8 +23,8 @@ dependencies {
 	testIntegrationImplementation project(":apps:portal-workflow:portal-workflow-kaleo-definition-api")
 	testIntegrationImplementation project(":apps:portal-workflow:portal-workflow-kaleo-runtime-api")
 	testIntegrationImplementation project(":apps:site:site-api")
-	testIntegrationImplementation project(":dxp:apps:ai-hub:ai-hub-api")
-	testIntegrationImplementation project(":dxp:apps:ai-hub:ai-hub-rest-api")
-	testIntegrationImplementation project(":dxp:apps:ai-hub:ai-hub-rest-client")
+	testIntegrationImplementation project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-api")
+	testIntegrationImplementation project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-rest-api")
+	testIntegrationImplementation project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-rest-client")
 	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
 }
